+++ src/Aneiang.Yarp.Plugin.Compression/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using Aneiang.Yarp.Plugins;
+global using Microsoft.AspNetCore.Builder;
+global using Microsoft.Extensions.DependencyInjection;